+++ changelog.md
@@ -1,3 +1,6 @@
+### [31-05-2020]
+* Rewrite /api/ ratelimit implementation to allow unlimited api for existed users(htpasswd), just set username and password as post arguments
+
 ### [17-05-2020]
 * [XLMMacroDeobfuscator](https://github.com/DissectMalware/XLMMacroDeobfuscator) from @DissectMalware integrated
 * Yara now compiled once at processing start or reprocessing